requirements.txt
fastapi==0.115.2

uvicorn[standard]==0.30.6

python-dotenv==1.0.1

requests==2.32.3

pydantic==2.9.2

numpy==1.26.4

soundfile==0.12.1

librosa==0.10.2.post1

torchaudio==2.3.1

torch==2.3.1

pyannote.audio==3.3.2

faster-whisper==1.0.3

tqdm==4.66.5

aiofiles==23.2.1